requirements.txt
# Data Processing
pandas>=2.0.0
duckdb>=0.9.0

# Airflow (optional, for orchestration)
apache-airflow>=2.7.0

# Testing
pytest>=7.0.0

# Utilities
python-dateutil>=2.8.0
